% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeA.R
\name{appendAnnotation}
\alias{appendAnnotation}
\title{--- old ---}
\usage{
appendAnnotation(
  gene.stat,
  ann,
  do.the.job = getOption("append.annot"),
  sort.by = 1
)
}
\arguments{
\item{gene.stat}{The table of genes, usually a DEG summary-statistic
top-table (or an expression matrix).}

\item{ann}{The matrix containing the annotation data.}

\item{do.the.job}{FALSE to skip the appending task by global settings,
without the need for an external IF.}

\item{sort.by}{The name or index of the column used to sort the final dataset}
}
\value{
The annotated and sorted data frame passed as input.
}
\description{
ppend annotation to genes and sort (do nothing if
do.the.job == FALSE).
}
\author{
\subsection{//--FeA.R--//}{
}
}
