% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeA.R
\name{array_create_annot}
\alias{array_create_annot}
\title{Title here}
\usage{
array_create_annot(platform, collapsing = FALSE)
}
\arguments{
\item{platform}{Affymetrix/Agilent platform annotation database.}

\item{collapsing}{Boolean flag to choose whether to collapse by unique Probe
ID in the case of annotation packages from Bioconductor
(GPL records from GEO are already collapsed).}
}
\value{
A data frame containing for each probe of the platform an number of
features selected by the user.
}
\description{
Create the annotation data frame starting from a database name
as returned by the \code{array_platform_selector()}. This function
implements a minimal graphical interface allowing the user to
select the number and the type of features to be used as
annotation.
}
\author{
\subsection{//--FeA.R--//}{
}
}
