% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeA.R
\name{array_platform_selector}
\alias{array_platform_selector}
\title{Title here}
\usage{
array_platform_selector(filt = "All")
}
\arguments{
\item{filt}{A string used to subset the list of platform annotations among
which to choose (i.e., a \code{grep} filter working on array names).
It can target array manufacturer ("Affymetrix", "Agilent"),
or the source of annotation ("Biocondutor", "GPL" for GEO).}
}
\value{
The name of the database corresponding to the platform chosen by the
user (to be used with \code{create.annot()} function).
}
\description{
A minimal graphical interface to retrieve the name of a suitable
annotation database (a \code{.db} R-package or a \code{GPL} GEO platform
record) starting from the selection of the name of a microarray
chip model. To be used only in interactive mode.
}
\examples{
\dontrun{
array_platform_selector() |> array_create_annot(platform) |> head()}
}
\author{
FeA.R
}
