% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeA.R
\name{basename2}
\alias{basename2}
\title{Get filename without extension.}
\usage{
basename2(file_name)
}
\arguments{
\item{file_name}{A string containing the file name to be \emph{basenamed}.}
}
\value{
A string containing the pure file name, without trailing path nor
extension.
}
\description{
Just like \code{basename()}, this wrapper of
\code{tools::file_path_sans_ext()} removes the path leading to the
file. However, unlike the native \code{basename()}, any file
extension is also removed.
}
\examples{
# Comparison with the native 'basename'
basename("/path/to/the/file.ext")
basename2("/path/to/the/file.ext")
}
\author{
FeA.R
}
