% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeA.R
\name{count_density}
\alias{count_density}
\title{Kernel Density Plot for RNA-Seq Counts}
\usage{
count_density(
  count_table,
  remove_zeros = TRUE,
  xlim = NULL,
  ylim = NULL,
  titles = c("Kernel Density Plot", deparse(substitute(count_table)))
)
}
\arguments{
\item{count_table}{A data frame containing gene expression counts. Only
numeric columns are allowed.}

\item{remove_zeros}{A Boolean flag. If set to \code{TRUE} (the default) zeros
are removed from count table beforehand.}

\item{xlim}{A two-element numeric vector specifying the visualization limits
for the x-axis. If set to \code{NULL} (the default), \code{plot()} default
values are used.}

\item{ylim}{A two-element numeric vector specifying the visualization limits
for the y-axis. If set to \code{NULL} (the default), \code{plot()} default
values are used.}

\item{titles}{Custom titles and subtitles for the Density Plot. A character
vector of two elements.}
}
\description{
Make a Kernel Density Plot from RNA-Seq data counts.
\code{count_density()} iterates \code{stats::density()} function over all
the columns of the \code{count_table} input data and produce a plot
with many density curves superimposed. In addition, it
implements a switch to remove zero counts (that usually prevent
a clear inspection of the data) and the possibility to set the
lower and upper bounds for both x- and y-axes (which is useful
for zooming in the region of interest).
}
\author{
FeA.R
}
