% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DEGs}
\alias{DEGs}
\title{Anti-TNFalpha-induced transcription in AR patients}
\format{
\subsection{\code{DEGs}}{

A data frame with 436 rows and 14 columns:
\describe{
\item{GeneSymbol}{HGNC Official Gene Symbol}
\item{GeneName}{Official Gene Name}
\item{Category}{Manual curated membership to one of the following
categories: NA, "Inflammation", "Immune Response", "Metallopeptidase"}
\item{Anti-TNFa_1 : Anti-TNFa_5}{anti-TNFalpha-treated patients}
\item{MTX_1 : MTX_6}{Methotrexate-treated patients}
}
}
}
\source{
PMID: ..., DOI: ..., GEO accession "GSE224330"
}
\usage{
DEGs
}
\description{
DEG list from my Rheumatoid Arthritis (AR) microarray study
(PMID: ..., DOI: ..., GEO accession "GSE224330").
This data set is a subset of the original one, in that it only contains the
expression values of statistically significant genes resulting from the sole
comparison of Etanercept (anti-TNFalpha) with Methotrexate (MTX)-treated
patients. Accordingly, only samples taking part in such a contrast (n = 5 for
Etanercept and n = 6 for MTX) have been included. Additional columns from 1
to 3 contain some gene annotation, namely: GeneSymbol, GeneName, and Category
membership. Row names are Agilent probe_IDs.
}
\author{
FeA.R
}
\keyword{datasets}
