% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DEGs_stat}
\alias{DEGs_stat}
\title{Anti-TNFalpha-induced transcription in AR patients - Statistics}
\format{
\subsection{\code{DEGs_stat}}{

A data frame with 436 rows and 5 columns:
\describe{
\item{\code{GENE_SYMBOL}}{HGNC Official Gene Symbol}
\item{\code{GENE_NAME}}{Official Gene Name}
\item{\code{logFC}}{log_2 Fold Changes of gene expression in anti-TNFalpha
compared to MTX-treated patients}
\item{\code{p_value}}{RankProduct-based (unadjusted) \emph{p}-values}
\item{\code{adj.p_value}}{Benjamini-Hochberg adjusted \emph{p}-values}
}
}
}
\source{
PMID: 36877690, DOI: 10.1371/journal.pone.0282564,
GEO accession "GSE224330".
}
\usage{
DEGs_stat
}
\description{
Descriptive (logFC) and inferential (\emph{p}-value) statistics for the same 436
significant genes already used to build the \code{DEGs_expr} data set and
resulting from the Differential Expression Analysis (DEA) of the
transcriptomics data from our 2017-2023 Rheumatoid Arthritis (AR) microarray
study (PMID: 36877690, DOI: 10.1371/journal.pone.0282564, GEO accession
"GSE224330").
Agilent-072363 SurePrint G3 Human GE v3 8x60K chips Chip were hybridized with
RNA of monocytes taken from bDMARD-treated AR-patients and compared with the
transcriptional profile of healthy and Methotrexate (MTX)-treated samples.
Raw florescence intensities were background-subtracted and interarray-
normalized to obtain the expression data matrix. DEA among the different
experimental groups were conducted using a RankProduct-based statistical
analysis. All genes with an adjusted \emph{p}-value < 0.05 and with a |FC| > 1.5
were deemed as statistically significant. The data set stored in this R
variable is a subset of the original one, in that it only contains the
statistics of the Differentially Expressed Genes (DEGs) resulting from the
sole comparison of Etanercept (anti-TNFalpha) with MTX-treated patients
(n = 5 and n = 6, respectively). Additionally, first two columns contain some
gene annotation, namely GeneSymbol and GeneName. Row names are Agilent
probe_IDs.
}
\author{
FeA.R
}
\keyword{datasets}
