% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeA.R
\name{descriptives}
\alias{descriptives}
\title{Basic Descriptive Statistics}
\usage{
descriptives(vals, design, labels = NULL, prec = 3)
}
\arguments{
\item{vals}{One-dimensional numeric vector or data frame.}

\item{design}{Experimental design: a numeric or character vector that
associates to each element of \code{vals} a symbol based on the
experimental group the element belong to. It can be shorter
than \code{vals}; in this case the last \code{length(vals)-length(design)}
elemnts of \code{vals} will be ignored (not included in any group).}

\item{labels}{Optional. A character vector containing the labels for the
experimental groups.}

\item{prec}{Decimal precision.}
}
\value{
A data frame containing the statistics of interest (currently sample
size, arithmetic mean, median, IQR, variance, standard deviation,
and SEM) for each group defined in \code{design}.
}
\description{
Use this function to get basics descriptive statistics of many
experimental groups from a single one-dimensional numeric vector
according to a user-defined experimental design. This function
is a generalization of the legacy \code{descStat1G()} function
implemented in GATTACA for single-gene inspection.
}
\author{
FeA.R
}
