% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeA.R
\name{descStat1G}
\alias{descStat1G}
\title{--- old ---}
\usage{
descStat1G(gene, gr, des, prec = 4)
}
\arguments{
\item{gene}{Numeric vector or single-row data frame from gene expression
matrix.}

\item{gr}{Group names.}

\item{des}{Experimental design (full design mode vector).}

\item{prec}{Decimal precision.}
}
\value{
A data frame containing the statistics of interest for each gene of
\code{gene}.
}
\description{
Return basics descriptive statistics of a single gene, by group
label. This is a GATTACA legacy piece of code.
}
\author{
FeA.R
}
