% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeA.R
\name{dnues2}
\alias{dnues2}
\title{Different Non-Unique Elements (Alternative)}
\usage{
dnues2(vec)
}
\arguments{
\item{vec}{A vector to be scanned for duplicates.}
}
\value{
A numeric vector of two elements, namely the number of different
non-unique elements, and the global number of non-unique entries in
\code{vec}.
}
\description{
An alternative (faster?) implementation of \code{dnues()},
independent of \code{dup_report()} function. A function that tells
how many \emph{different non-unique} elements there are in a given
vector. Unlike \code{base::duplicated()} that detects duplicated
entries \emph{after} their first occurrence, this function looks at
\emph{all} non-unique values. \code{NA}s are ignored by default.
}
\examples{
# How many different non-unique Gene Symbols are there in DEGs data set? 
dnues2(DEGs$GeneSymbol)
}
\author{
FeA.R
}
