% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeA.R
\name{duplication_report}
\alias{duplication_report}
\title{Title here}
\usage{
duplication_report(vec)
}
\arguments{
\item{vec}{A vector to be scanned for duplicates.}
}
\value{
A list of the duplicated elements found in \code{vec}. Each element of
this list is in turn a list of two elements: an integer indicating
the times the element is repeated and a vector of integers
representing its positions within \code{vec}.
}
\description{
This function searches the vector passed as input for duplicated
entries and the positions they fill in the vector. Unlike
\code{base::duplicated()} that only detects duplicated entries
\emph{after} their first occurrence, this function looks at \emph{all}
non-unique values.
}
\author{
\subsection{//--FeA.R--//}{
}
}
