% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeA.R
\name{echo}
\alias{echo}
\title{Colored echo for on-screen log in Bash CLI}
\usage{
echo(text, color = "white")
}
\arguments{
\item{text}{A string containing the text to be printed in console.}

\item{color}{One of the following strings: "white" (the default), "red",
"green", "yellow", "blue", "magenta", "cyan".}
}
\description{
A wrapper for easy managing color output in Linux terminal.
}
\examples{
\dontrun{
# From Bash CLI
Rscript -e 'r4tcpl::echo("GOOD STUFF", "green")'
}
}
\author{
FeA.R
}
