% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeA.R
\name{fpkm2tpm}
\alias{fpkm2tpm}
\title{FPKM to TPM}
\usage{
fpkm2tpm(fpkm)
}
\arguments{
\item{fpkm}{A data frame containing unlogged normalized counts in FPKM units.
Only numeric columns are allowed.}
}
\value{
A data frame containing unlogged normalized counts in TPM units.
}
\description{
A function to convert normalized RNA-Seq counts from FPKM (or
RPKM) to TPM scale. \strong{NOTE:} to be properly converted, the
input FPKM counts need to be linearly scaled
(\strong{NOT log-transformed!})
}
\author{
FeA.R
}
