% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeA.R
\name{howMany_dnues}
\alias{howMany_dnues}
\title{Different Non-Unique Elements}
\usage{
howMany_dnues(vec)
}
\arguments{
\item{vec}{A vector to be scanned for duplicates.}
}
\value{
A vector of two elements, namely the number of different non-unique
elements, and the global number of non-unique entries in \code{vec}.
}
\description{
A function that tells how many \emph{different non-unique} elements
there are in a given vector. Unlike \code{base::duplicated()} that
detects duplicated entries \emph{after} their first occurrence, this
function looks at \emph{all} non-unique values.
}
\author{
FeA.R
}
