% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeA.R
\name{howMany_dnues2}
\alias{howMany_dnues2}
\title{Title here}
\usage{
howMany_dnues2(vec)
}
\arguments{
\item{vec}{A vector to be scanned for duplicates.}
}
\value{
A vector of two elements, namely the number of different non-unique
elements, and the global number of non-unique entries in \code{vec}.
}
\description{
An alternative (faster?) implementation of \code{howMany_dnues()},
independent of \code{duplication_report()}.
}
\author{
\subsection{//--FeA.R--//}{
}
}
