% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ion_tools.R
\name{iMap}
\alias{iMap}
\title{Current Map}
\usage{
iMap(dataset)
}
\arguments{
\item{dataset}{A data frame or a matrix containing a different single-channel
current trace \strong{per row}.}
}
\description{
This function provides a comprehensive graphical representation
of many single-channel ion currents as a 2D colormap.
}
\author{
FeA.R
}
