% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ISO}
\alias{ISO}
\title{Titanium surface descriptors (ISO 25178)}
\format{
\subsection{\code{ISO}}{

A data frame with 8 rows and 11 columns:
\describe{
\item{\code{Sa}}{Arithmetical mean height of the surface}
\item{\code{Sku}}{Kurtosis of height distribution}
\item{\code{Sp}}{Maximum height of peaks}
\item{\code{Sq}}{Root mean square height of the surface}
\item{\code{Ssk}}{Skewness of height distribution}
\item{\code{Sv}}{Maximum height of valleys}
\item{\code{Sz}}{	Maximum height of the surface}
\item{\code{Sdq}}{Root mean square gradient of the surface}
\item{\code{Sdr}}{Developed area ratio}
\item{\code{Sal}}{Fastest decay auto-correlation rate}
\item{\code{Str}}{Texture aspect ratio of the surface}
}
}
}
\source{
PMID: ..., DOI: ..., \url{https://en.wikipedia.org/wiki/ISO_25178}
}
\usage{
ISO
}
\description{
This data set contains the experimentally measured values of eleven ISO 25178
(Geometrical Product Specifications, GPS, Surface texture: areal) descriptors
evaluated for eight different types of titanium disks. These data, together
with the related profiles of cell adhesion and protein adsorption, have been
presented, analyzed, and discussed in the following paper: PMID: ...,
DOI: ...,
}
\author{
FeA.R
}
\keyword{datasets}
