% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ion_tools.R
\name{iTrace}
\alias{iTrace}
\title{Current Plot}
\usage{
iTrace(dataset)
}
\arguments{
\item{dataset}{A data frame or a matrix containing a different single-channel
current trace \strong{per row}.}
}
\description{
A multi-trace viewer for single-channel ion currents. Starting
from a data set containing many current traces, \code{iTrace()} makes
a composite plot showing the first three single-channel traces
and a fourth subplot with the total current. This is meant to
highlight the transition from the stochastic behavior of the
single ion channel to the deterministic electrical current
observed at the whole-cell level.
}
\author{
FeA.R
}
