% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeA.R
\name{quick_chart}
\alias{quick_chart}
\title{Standard plots for categorical data}
\usage{
quick_chart(vals, design, chart_type = "BP")
}
\arguments{
\item{vals}{One-dimensional numeric vector or data frame.}

\item{design}{Experimental design: a numeric or character vector that
associates each element of \code{vals} to a symbol, based on the
experimental group the element belongs to.}

\item{chart_type}{A string among the following: "BP" (Box Plot), "VP"
(Violin Plot), "BC" (Bar Chart), or "MS" (Mean & SEM).}
}
\description{
Use this function to plot a continuous response variable as a
function of a categorical explanatory one. Data need to be
passed as a single one-dimensional numeric vector together with
a user-defined experimental design. This function is a
generalization of the legacy \code{singleGeneView()} function
implemented in GATTACA for single-gene inspection.
}
\author{
FeA.R
}
