% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeA.R
\name{relog}
\alias{relog}
\title{Pseudocounted log function for RNA-Seq data}
\usage{
relog(x)
}
\arguments{
\item{x}{A numeric (usually integer) single value, matrix, or data frame.}
}
\value{
The log counts.
}
\description{
Takes the base-2 logarithm with a pseudo count of 1.
}
\author{
FeA.R
}
