% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeA.R
\name{repmat}
\alias{repmat}
\title{Title here}
\usage{
repmat(X, m, n)
}
\arguments{
\item{X}{A vector, matrix, or data frame with source array data.}

\item{m}{An integer specifying the number of row-wise repetitions.}

\item{n}{An integer specifying the number of column-wise repetitions.}
}
\value{
A new matrix resulting from the m-by-n juxtaposition of the starting
matrix \code{X}.
}
\description{
R-equivalent to the MATLAB \code{repmat()} function. Just specify the
matrix X and how many times you want it replicated row- (m) and
column- (n) wise.
}
\author{
FeA.R
}
