% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeA.R
\name{rmv}
\alias{rmv}
\title{Remove substring}
\usage{
rmv(x, rgx)
}
\arguments{
\item{x}{The input string.}

\item{rgx}{The regex pattern to be found and removed.}
}
\description{
A simple wrapper for \code{gsub()} to remove \strong{all} the substrings
that match a given regex pattern. Notably, \code{rmv()} is compliant
with the native pipe operator.
}
\examples{
# Make them singular
"children, cars, and witches" |> rmv("e?s|ren")
# Remove leading and trailing spaces
"   Let's clean this !   " |> rmv("^( +)") |> rmv("( +)$")
# Convert an array from Bash to an R character vector
"(foo bar this that)" |> rmv("\\\\(|\\\\)") |> strsplit(" +") |> unlist()
}
\author{
FeA.R
}
