% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeA.R
\name{rotate}
\alias{rotate}
\title{Rotate a matrix or data frame}
\usage{
rotate(mat, turns = 1)
}
\arguments{
\item{mat}{A matrix, or data frame. Row- and column-names are handled.}

\item{turns}{Number of 90° clockwise rotations. Use negative integers for
counterclockwise rotations.}
}
\value{
A rotated matrix. If \code{mat} is a data frame, it will be converted to
matrix during transposition.
}
\description{
This function takes in a matrix or a data frame and rotates it
a user-defined number of times.
}
\details{
R modulus definition is consistent with a "floored division"
approach, so that \code{((a \%/\% b)*b + a \%\% b) == a}. For this reason,
taking the modulus 4 directly implements a counterclockwise rotation
when the operand is negative!
}
\examples{
# Clockwise rotation
rotate(nanto$scores)
# Counterclockwise rotation
rotate(nanto$scores, -3)
}
\author{
FeA.R
}
