% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeA.R
\name{savePlots}
\alias{savePlots}
\title{Save plot to a file in multiple formats}
\usage{
savePlots(
  plotfun,
  ratio = 16/9,
  figure_Name,
  figure_Folder,
  png_out = TRUE,
  pdf_out = TRUE
)
}
\arguments{
\item{plotfun}{A callback function that prints a plot to an open device.}

\item{ratio}{The plot aspect ratio, as a single numeric input.}

\item{figure_Name}{Name of the output file (without extension).}

\item{figure_Folder}{Name of the saving (sub)folder.}

\item{png_out}{Boolean. Set it to \code{FALSE} to suppress PNG graphics device.}

\item{pdf_out}{Boolean. Set it to \code{FALSE} to suppress PDF graphics device.}
}
\description{
This function saves a graphical output to \code{figure_Folder}
sub-directory in both raster (PNG) and vectorial (PDF) formats.
Automatically makes the output folder if not already there.
}
\details{
This implementation of \code{savePlots()} takes a \emph{function} as its first
argument, and not a plot object (i.e., a device number) as some
plotting facilities (notably the default one, \code{savePlot()}) that
cannot print plot objects conveniently.
}
\examples{
\dontrun{
# Get a numeric-only expression matrix (as a data frame)
numeric_df <- DEGs_expr[,4:14]

# Save box plots of per-sample distributions of expression levels
savePlots(
  \(){boxplot(numeric_df)},
  figure_Name = "Boxplot",
  figure_Folder = ".")

# Calculate PCA on samples
metadata <- data.frame(row.names = colnames(numeric_df))
metadata$Group <- c(rep("AntiTNF",5), rep("MTX",6))
pcaOut <- PCAtools::pca(numeric_df, metadata = metadata)

# Print the biplot
savePlots(
  \(){print(PCAtools::biplot(pcaOut, colby = "Group"))},
  figure_Name = "BiPlot",
  figure_Folder = ".")
}
}
\author{
FeA.R, Hedmad
}
