% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeA.R
\name{seq_GMM}
\alias{seq_GMM}
\title{Gaussian Mixture Model applied to RNA-Seq count data}
\usage{
seq_GMM(count_data, comp_num = 2, sub_pops = c(1, comp_num))
}
\arguments{
\item{count_data}{One-dimensional numeric vector or data frame.}

\item{comp_num}{Number of gaussian components to be used for the mixture.}

\item{sub_pops}{A two-element integer vector containing the index of the two
components to intersect to find the boundary.}
}
\value{
A named list with the following elements:
\describe{
\item{fit}{An object of class 'Mclust' providing the GMM estimation.}
\item{x}{A 1000-point numeric vector providing the x-values used
to evaluate the components. Useful for plotting.}
\item{components}{A 1000 x \code{comp_num} data frame featuring the
probability-weighted Gaussian components evaluated in x.}
\item{boundary}{The decision boundary computed as the intersection between
the two Gaussian components specified by the \code{sub_pops}
argument.}
}
}
\description{
Use a Gaussian Mixture Model (GMM) to fit RNA-Seq count
distribution in order to identify expression sub-populations
(i.e., expressed vs unexpressed genes). Starting from unbinned
counts, this function returns the individual Gaussian components
of the mixture and a boundary value suitable for sub-population
separation.
}
\examples{
log_expression <- DEGs_expr$`Anti-TNFa_4`
plot(density(DEGs_expr$`Anti-TNFa_4`), main = "Kernel Density Plot")

gmm <- seq_GMM(log_expression, comp_num = 3, sub_pops = c(1,2))

for (i in 1:gmm$fit$G) {
  lines(gmm$x, gmm$components[,i], col = "blue")
}
lines(gmm$x, rowSums(gmm$components), col = "red", lty = 2)
lines(c(gmm$boundary, gmm$boundary), c(0, 1), lty = 2)
}
\author{
FeA.R
}
