% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_exploration.R
\name{show.data}
\alias{show.data}
\title{Show a snippet of the dataset}
\usage{
show.data(dataset, name = NULL, rows = 10, cols = 5)
}
\arguments{
\item{dataset}{A \code{data.frame} to show}

\item{name}{The name of the dataset, used to enrich the output.}

\item{rows}{The number of rows to show.}

\item{cols}{The number of cols to show.}
}
\value{
A dimesion vector of the form \code{c(rows, cols)}.
}
\description{
This function is superseeded by the \code{print} implementation of a \code{tibble}.
It is included here as an example function.
}
\examples{
x <- data.frame(var1 = c(1, 2, 3, 4, 5), var2 = c(6, 7, 8, 9, 10))
show.data(x, "an example dataset", rows = 3)
}
