% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeA.R
\name{show_data}
\alias{show_data}
\title{Show Data as matrix and return its dimensions}
\usage{
show_data(dataset, name = NULL, rows = 10, cols = 5)
}
\arguments{
\item{dataset}{Data frame, matrix, or vector to print.}

\item{name}{Explanatory name to print in the heading (useful when logging).}

\item{rows}{Maximum number of rows to display.}

\item{cols}{Maximum number of columns to display.}
}
\value{
A vector of the form \code{c(rows,cols)} containing \code{dataset} dimensions.
}
\description{
A custom version of the classical \code{head()} that prints the upper
leftmost corner of a data set, also showing row names and
controlling for possible out-of-bounds exceptions. Compared to
\code{head()}, \code{show_data()} \strong{always displays data by columns},
even in the case of vectors (i.e., one-dimensional arrays).
Also, \code{show_data} prints and returns the dimensions of the data
set, together with a custom heading label.
}
\examples{
\dontrun{
x <- data.frame(var1 = c(1, 2, 3, 4, 5), var2 = c(6, 7, 8, 9, 10))
show_data(x, "an example dataset", rows = 3)}
}
\author{
FeA.R
}
