% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeA.R
\name{singleGeneView}
\alias{singleGeneView}
\title{--- old ---}
\usage{
singleGeneView(exp.mat, gr, des, gois, chart.type = "BP", ann = NULL)
}
\arguments{
\item{exp.mat}{Expression matrix (as data frame).}

\item{gr}{Group names.}

\item{des}{Experimental design (full design mode vector).}

\item{gois}{Genes of interest by probe (char vector).}

\item{chart.type}{"BP" (Box Plot), "BC" (Bar Chart), or "MS" (Mean & SEM).}

\item{ann}{Optional annotation data frame.}
}
\description{
Plot single gene comparison chart. This is a GATTACA legacy
piece of code.
}
\author{
FeA.R
}
