% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeA.R
\name{tab}
\alias{tab}
\title{Tab Stop}
\usage{
tab(word = "", sp = 7)
}
\arguments{
\item{word}{The string to be included in the tabular space.}

\item{sp}{Spacing parameter indicating tab stop position (i.e., the total
width of the final string returned by the function).}
}
\value{
A string consisting of the \code{word} followed by a variable number of
white spaces.
}
\description{
To align strings in console as if using MS-Word tab stops
It allows to control tab stop positions padding the given
string with white spaces to reach a fixed width.
}
\examples{
# Print Nanto warriors' hair color
for (warrior in nanto$warrior) {
  cat(" -", tab(warrior), ":: has",
      tab(nanto$all_data[warrior, "hair_color"], 8), "hair\n")
}
}
\author{
FeA.R
}
