% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TGS}
\alias{TGS}
\title{The Transportome Gene Set - TGS}
\format{
\subsection{\code{TGS}}{

A named list containing 5 character vectors:
\describe{
\item{\code{pseudo}}{All human ICTs, including pseudogenes}
\item{\code{whole}}{All human ICTs, without pseudogenes (a subset of \code{pseudo})}
\item{\code{ICs}}{All human Ion Channels (a subset of \code{whole})}
\item{\code{Ca_ICs}}{All human Calcium-Permeable Ion Channels (a subset of \code{ICs})}
\item{\code{trans}}{All human transporters, including solute carriers (SLCs),
ABC transporters, and ATPase pumps (a subset of \code{whole} with
null overlap with \code{ICs}).}
}
}
}
\source{
HGNC DB at \url{https://www.genenames.org/} and
IUPHAR/BPS DB at \url{https://www.guidetopharmacology.org/}
}
\usage{
TGS
}
\description{
This is a comprehensive list of the gene symbols representing the entire
human transportome, defined as the set of all the Ion Channels and
Transporters (ICTs) that can be found across the human cell membrane. Gene
Symbols are organized in 5 character vectors as described below.
}
\author{
FeA.R
}
\keyword{datasets}
