% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeA.R
\name{unlog}
\alias{unlog}
\title{Pseudocounted log^(-1) function for RNA-Seq data}
\usage{
unlog(x)
}
\arguments{
\item{x}{A numeric single value, matrix, or data frame.}
}
\value{
The unlogged counts.
}
\description{
Takes back to linear scale, accounting for a pseudo count of 1.
}
\author{
FeA.R
}
