% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeA.R
\name{venny}
\alias{venny}
\title{Overlap Analysis}
\usage{
venny(
  set_A,
  set_B,
  N = 20000,
  venn = TRUE,
  lab = c("Set A", "Set B"),
  titles = c("Venn diagram", "by cmatools")
)
}
\arguments{
\item{set_A}{A character or numeric (1D) vector.}

\item{set_B}{Another character or numeric (1D) vector.}

\item{N}{Size of the universe (or background set). see \code{hgt()} function.}

\item{venn}{Boolean. Set it to \code{FALSE} to suppress Venn plotting.}

\item{lab}{Labels for the Venn. A character vector of two elements.}

\item{titles}{Titles and subtitles for the Venn diagram. A character vector
of two elements.}
}
\value{
A list made up of a data frame named \code{ORA} (containing the results
of the OverRepresentation Analysis), and three vectors named
\code{intersection}, \code{left_set}, and \code{right_set}, featuring the elements
of those three sets, respectively.
}
\description{
Given \strong{two} vectors of symbols, this function finds the
elements that exist in both sets and computes the statistical
significance of the overlap between them. Optionally, it plots
a Venn diagram representation.
}
\examples{
# Find how many (and which) Ion Channels there are within a given DEG list
# and if that gene set is enriched (over-represented) or not:
x <- venny(TGS$ICs, DEGs_stat$GENE_SYMBOL,
           lab = c("Ion Channels","DEGs"),
           titles = c("Transportome Analysis", "Ion Channels"))
}
\author{
FeA.R
}
