% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{x}
\alias{x}
\title{General Test Variable (Animals, Ages, and Owners)}
\format{
\subsection{\code{x}}{

A list containing 12 named objects of different types.
\describe{
\item{animal}{Character vector with no duplicated entries}
\item{owner}{Character vector with duplicated entries}
\item{named_animal}{Named character vector (with duplicated names)}
\item{a_age}{Numeric vector with no duplicated entries}
\item{o_age}{Numeric vector with duplicated entries}
\item{named_ages}{Named numeric vector (with duplicated names)}
\item{fur}{Logical vector}
\item{sex}{A factor with 3 levels}
\item{whos}{Character matrix}
\item{ages}{Numeric matrix}
\item{named_ages}{Numeric matrix with (duplicated) row and column names}
\item{all_data}{A data frame containing all the previous objects}
}
}
}
\source{
Created by me (FeA.R). See comments in .../cmatools/R/data.R to
recreate or modify the dataset.
}
\usage{
x
}
\description{
A list containing the most common basic data type you can find in R. Useful
for interactive trials and testing.
}
\keyword{datasets}
