% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeA.R
\name{array_create_annot}
\alias{array_create_annot}
\title{Microarray Annotation Retriever}
\usage{
array_create_annot(platform, collapsing = FALSE)
}
\arguments{
\item{platform}{Affymetrix/Agilent platform annotation database as returned
by the \code{array_platform_selector()} function.}

\item{collapsing}{Boolean flag to choose whether to collapse by unique Probe
ID in the case of annotation packages from Bioconductor
(GPL records from GEO are already collapsed).}
}
\value{
A data frame containing for each probe of the platform a number of
features as selected by the user.
}
\description{
This function retrieves gene annotation for a given microarray
platform and returns them as a data frame. As input, it requires
a database name as returned by \code{array_platform_selector()}.
This function uses \code{svDialogs} package to implement a minimal
graphical interface allowing the user to select the number and
the type of features to be used as annotation. GPL-based
annotation are retrieved from GEO using \code{GEOquery::getGEO()}
function that needs a working FTP connection.
}
\examples{
\dontrun{
annot <- array_create_annot("hgu133plus2", collapsing = TRUE)
missing_report(annot)
lms(annot)}
}
\author{
FeA.R
}
