% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DEGs}
\alias{DEGs}
\title{Anti-TNFalpha-induced differential gene expression in AR patients.}
\format{
\subsection{\code{DEGs}}{

A data frame with 436 rows and 14 columns:
\describe{
\item{GeneSymbol}{HGNC Official Gene Symbol}
\item{GeneName}{Official Gene Name}
\item{Category}{Manual curated membership to one of the following
categories: NA, "Inflammation", "Immune Response", "Metallopeptidase"}
\item{Anti-TNFa_1 : Anti-TNFa_5}{anti-TNFalpha-treated patients}
\item{MTX_1 : MTX_6}{Methotrexate-treated patients}
}
}
}
\source{
\url{https://www.referenzaweb}
}
\usage{
DEGs
}
\description{
DEG list from my Rheumatoid Arthritis (AR) microarray study (references...)
This subset contains the expression values for each sample taking part in the
Etanercept (anti-TNFalpha) vs Methotrexate (MTX) contrast (n = 5 and n = 6,
respectively). Row names are Agilent probe_IDs. Additional columns (1 to 3)
contains some gene annotation, namely: GeneSymbol, GeneName, and Category
membership.
Row-wise, only statistically significant genes are included in this dataset.
}
\keyword{datasets}
