% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeA.R
\name{descriptives}
\alias{descriptives}
\title{Basic Descriptive Statistics}
\usage{
descriptives(vals, design = rep(1, length(vals)), prec = 3)
}
\arguments{
\item{vals}{One-dimensional numeric vector or data frame.}

\item{design}{Experimental design: a numeric or character vector that
associates each element of \code{vals} to a symbol, based on the
experimental group the element belongs to.}

\item{prec}{Output decimal precision.}
}
\value{
A data frame containing the statistics of interest (i.e., sample
size, arithmetic mean, median, IQR, variance, standard deviation,
and SEM) for each group defined in \code{design}.
}
\description{
Use this function to get basic descriptive statistics of many
experimental groups from a single one-dimensional numeric vector
according to a user-defined experimental design. This function
is a generalization of the legacy \code{descStat1G()} function
implemented in GATTACA for single-gene inspection.
}
\examples{
# Get descriptive statistics of gene expression
gene <- DEGs["A_33_P3307955", -c(1:3)]
dsgn <- c(rep("Anti-TNFa",5), rep("MTX",6))
descriptives(gene, dsgn, 4)
}
\author{
FeA.R
}
