% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeA.R
\name{hgt}
\alias{hgt}
\title{Title here}
\usage{
hgt(k, n, K, N = 10000)
}
\arguments{
\item{k}{Number of hits in the experimental set.  --OR--> Intersection}

\item{n}{Size of the experimental set.            --OR--> Set A}

\item{K}{Number of possible hits in the universe. --OR--> Set B}

\item{N}{Size of the universe.                    --OR--> Background}
}
\value{
A 1-by-5 data frame containing enrichment statistics. Many
single-row data frames can be easily stacked using \code{rbind()}.
}
\description{
A wrapper for the Hypergeometric test function. Beside
\emph{p}-values, \code{hgt()} also computes and returns other enrichment
statistics like the expected value and fold enrichment.
}
\author{
FeA.R
}
