% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ISO}
\alias{ISO}
\title{Titanium surface ISO 25178 descriptors}
\format{
\subsection{\code{ISO}}{

A data frame with 8 rows and 11 columns:
\describe{
\item{Sa}{Arithmetical mean height of the surface}
\item{Sku}{Kurtosis of height distribution}
\item{Sp}{Maximum height of peaks}
\item{Sq}{Root mean square height of the surface}
\item{Ssk}{Skewness of height distribution}
\item{Sv}{Maximum height of valleys}
\item{Sz}{	Maximum height of the surface}
\item{Sdq}{Root mean square gradient of the surface}
\item{Sdr}{Developed area ratio}
\item{Sal}{Fastest decay auto-correlation rate}
\item{Str}{Texture aspect ratio of the surface}
}
}
}
\source{
\url{https://en.wikipedia.org/wiki/ISO_25178}
}
\usage{
ISO
}
\description{
ISO 25178: Geometrical Product Specifications (GPS) – Surface texture: areal.
}
\keyword{datasets}
