% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeA.R
\name{lms}
\alias{lms}
\title{Let Me See (Data as Matrix)}
\usage{
lms(data2see, rows = 10, cols = 5, name = NULL)
}
\arguments{
\item{data2see}{Data frame, matrix, factor, or vector to print.}

\item{rows}{Maximum number of rows to display.}

\item{cols}{Maximum number of columns to display.}

\item{name}{Explanatory name to print in the heading (useful when logging).}
}
\description{
A custom version of the classical \code{head()} that prints the upper
leftmost corner of a data set, also showing row names and
controlling for possible out-of-bounds exceptions. Compared to
\code{head()}, \code{lms()} \strong{displays 1D vectors by columns}, unquotes
strings in character vectors and matrices, and prints the
dimensions and the type (\code{class()}) of the data set, along with
a custom heading label. In addition, \code{lms()} features a better
list management.
}
\examples{
# Let me see The Nanto Warriors Data Set
lms(nanto)

# Compare `head()` behavior when applied to a list of objects
print(head(nanto))

# Compare `lms()` and `head()` behavior when applied to different data types
for (i in 1:length(nanto)) {
  cat("\nhead()\n")
  print(head(nanto[[i]]))
  cat("\nlms()\n")
  lms(nanto[[i]])
}
}
\author{
FeA.R
}
