% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeA.R
\name{lms}
\alias{lms}
\title{Let Me See (Data as Matrix)}
\usage{
lms(dataset, name = NULL, rows = 10, cols = 5)
}
\arguments{
\item{dataset}{Data frame, matrix, or vector to print.}

\item{name}{Explanatory name to print in the heading (useful when logging).}

\item{rows}{Maximum number of rows to display.}

\item{cols}{Maximum number of columns to display.}
}
\description{
A custom version of the classical \code{head()} that prints the upper
leftmost corner of a data set, also showing row names and
controlling for possible out-of-bounds exceptions. Compared to
\code{head()}, \code{lms()} \strong{always displays data by columns}, even in
the case of vectors (i.e., one-dimensional arrays). Also, \code{lms}
prints the dimensions of the data set, along with a custom
heading label.
}
\examples{
\dontrun{
x <- data.frame(var1 = c(1, 2, 3, 4, 5), var2 = c(6, 7, 8, 9, 10))
lms(x, "an example dataset", rows = 3)}
}
\author{
FeA.R
}
