% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeA.R
\name{missing_report}
\alias{missing_report}
\title{Count Missing Values in a Matrix-like Object}
\usage{
missing_report(dataFrame, naSymb = "")
}
\arguments{
\item{dataFrame}{Data frame or matrix to be scanned for NAs.}

\item{naSymb}{A string containing the user-defined sequence for NAs.}
}
\description{
This function takes a data frame, searches its columns for many
common missing-value placeholders, and finally prints a report
report of the number and the relative amount of missing values
detected column-wise. In particular, it searches for \code{NA},
\code{"NA"}, and patterns of whitespaces (\verb{\\s}), hyphens (\code{-}),
slashes (\code{/}), including empty fields and a possible
user-defined sequence.
}
\examples{
\dontrun{
array_platform_selector() |> array_create_annot() |> missing_report()}
}
\author{
FeA.R
}
