% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nanto}
\alias{nanto}
\title{The Nanto Warriors Data Set}
\format{
\subsection{\code{nanto}}{

A named list containing 12 objects of different types:
\describe{
\item{warrior}{Character vector with no duplicated entries}
\item{hair_color}{Character vector with duplicated entries}
\item{named_warrior}{Named character vector (with duplicated names)}
\item{power}{Numeric (float) vector with no duplicated entries}
\item{w_age}{Numeric (integer) vector with duplicated entries and NAs}
\item{named_pwr}{Named numeric vector}
\item{killed_by_Ken}{Logical vector}
\item{sex}{A factor with 3 levels}
\item{whos}{Character matrix}
\item{scores}{Numeric matrix}
\item{named_scores}{Numeric matrix with row and column names}
\item{all_data}{A data frame containing all the previous objects}
}
}
}
\source{
\url{https://en.wikipedia.org/wiki/List_of_Fist_of_the_North_Star_characters}
}
\usage{
nanto
}
\description{
This data set, inspired by the six Nanto warriors of the epic
\strong{Hokuto no Ken} (aka \emph{Fist of the North Star}) saga, is provided in the
form of a named list containing some of the most common basic data types you
can find in R. It is very useful for educational purposes, interactive
trials, and general testing.
}
\author{
FeA.R
}
\keyword{datasets}
