% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeA.R
\name{quick_chart}
\alias{quick_chart}
\title{Standard Plots for Categorical Data}
\usage{
quick_chart(vals, design, chart_type = "BP")
}
\arguments{
\item{vals}{One-dimensional numeric vector or data frame.}

\item{design}{Experimental design: a numeric or character vector that
associates each element of \code{vals} to a symbol, based on the
experimental group the element belongs to.}

\item{chart_type}{A string among the following: "BP" (Box Plot), "VP"
(Violin Plot), "BC" (Bar Chart), or "MS" (Mean & SEM).}
}
\description{
Use this function to plot a continuous response variable as a
function of a categorical explanatory one. Data need to be
passed as a single one-dimensional numeric vector together with
a user-defined experimental design. This function is a
generalization of the legacy \code{singleGeneView()} function
implemented in GATTACA for single-gene inspection.
}
\examples{
# Get a graphical representation of gene differential expression
gene <- DEGs_expr["A_33_P3307955", -c(1:3)]
dsgn <- c(rep("Anti-TNFa",5), rep("MTX",6))
for (type in c("BP", "VP", "BC", "MS")) {
  quick_chart(gene, dsgn, type)
}
}
\author{
FeA.R
}
