% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeA.R
\name{seq_GMM}
\alias{seq_GMM}
\title{Gaussian Mixture Model applied to RNA-Seq count data}
\usage{
seq_GMM(count_data, comp_num = 2, sub_pops = c(1, comp_num))
}
\arguments{
\item{count_data}{One-dimensional numeric vector or data frame.}

\item{comp_num}{Number of components to be used for the mixture.}

\item{sub_pops}{A two-element integer vector containing the index of the two
components to intersect to find the boundary.}
}
\value{
A list with the following elements:
}
\description{
Use a Gaussian Mixture Model (GMM) to fit RNA-Seq count
distribution in order to identify the sub-populations of
expressed and unexpressed genes. Starting from raw (unbinned)
data, this function returns the individual Gaussian components
of the mixture and a boundary value suitable for subpopulation
separation.
}
\examples{
\dontrun{
log_expression <- DEGs_expr$`Anti-TNFa_4`
plot(density(DEGs_expr$`Anti-TNFa_4`), main = "Kernel Density Plot")

gmm <- seq_GMM(log_expression, comp_num = 3, sub_pops = c(1,2))

for (i in 1:gmm$fit$G) {
  lines(gmm$x, gmm$components[,i], col = "blue")
}
lines(gmm$x, rowSums(gmm$components), col = "red", lty = 2)
y_lim <- par("yaxp")[2]
lines(c(gmm$boundary, gmm$boundary), c(0, 1.5*y_lim), lty = 2)
}
}
\author{
FeA.R
}
