% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeA.R
\name{tab}
\alias{tab}
\title{Title here}
\usage{
tab(word = "", sp = 7)
}
\arguments{
\item{word}{The string to be included in the tabular space.}

\item{sp}{Spacing parameter indicating tab stop position (i.e., the total
width of the final string returned by the function).}
}
\value{
A string consisting of the \code{word} followed by a variable number of
white spaces.
}
\description{
To align strings in console as if using MS-Word tab stop
feature. It allows to control tab stop positions filling the
given string with white spaces to reach a fixed width.
}
\examples{
\dontrun{
for (animal in c("dog", "sheep", "worm", "lioness", "monkey")) {
   cat(" -", tab(animal), ":: looking for a", animal, "...found!\n")}}
}
\author{
FeA.R
}
